% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P01_IDDConf-package.R
\docType{package}
\name{P01_IDDConf-package}
\alias{P01_IDDConf}
\alias{P01_IDDConf-package}
\title{P01_IDDConf: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lauraguzmanrincon/P01_IDDConf}
  \item Report bugs at \url{https://github.com/lauraguzmanrincon/P01_IDDConf/issues}
}

}
\author{
\strong{Maintainer}: Laura Guzman-Rincon  \email{lauramgr17@gmail.com} (\href{https://orcid.org/0000-0003-4592-2896}{ORCID}) [copyright holder]

}
\keyword{internal}
